dns: always publish records to configured DNS zones
Before this change, DNS records were only published to the configured public DNS
if the ingresscontroller load balancer is public. After this change, always
publish DNS records to any configured private or public DNS zone regardless of
the ingresscontroller load balancer scope. The new default seems more generally
useful, especially in the contexts in which internal load balancers are used.
Non-cluster clients on the same VPC don't always resolve DNS through the
internal zone, and so need to resolve the load balancer's internal hostname
through a public zone.

This change also seems more consistent with our published DNS configuration API
which doesn't mention the current behavior with regards to DNS and
internal/external. In fact, it's clear in the DNS configuration docs that the
presence of a public or private zone is the way to configure whether records
should be published in a given zone.

One notable implication of this change is that the public zone DNS record
pointing at the internal load balancer leaks a detail about the internal load
balancer (it hostname or IP).