ensureServiceMonitor: Update resource if needed
If a servicemonitor already exists but is not the same as the desired
servicemonitor, update the current servicemonitor.

Commit 0adb643befab1e435c53aefd7c16e0a84dba5fe5 added a selector to the
servicemonitor that desiredServiceMonitor returns, so the operator must
update any servicemonitors created by older versions of the operator.

This commit fixes bug 1740595.

https://bugzilla.redhat.com/show_bug.cgi?id=1740595

* manifests/00-cluster-role.yaml: Allow updating servicemonitors.
* pkg/operator/controller/ingress/monitoring.go (ensureServiceMonitor): Add
godoc.  If a servicemonitor is needed and none exists, use
createServiceMonitor to create it.  If a servicemonitor already exists,
update it if needed using updateServiceMonitor.  Finally, return a Boolean
value indicating whether a servicemonitor exists.
(desiredServiceMonitor): Add godoc.  Fix the type of the "endpoints" field,
and add a comment why it is important to use the right type.
(currentServiceMonitor): Add godoc.  Add a Boolean return value indicating
whether a servicemonitor exists.
(createServiceMonitor, updateServiceMonitor): New methods, used by
ensureServiceMonitor.
(serviceMonitorChanged): New function, used by updateServiceMonitor.
Compare the specs of the given servicemonitors and return a Boolean value
indicating whether an update is needed and, if so, an updated
servicemonitor definition.
* pkg/operator/controller/ingress/metrics.go (ensureMetricsIntegration):
Ignore the Boolean return value from ensureServiceMonitor.
* pkg/operator/controller/ingress/monitoring_test.go
(TestServiceMonitorChanged): New test.