Bug 1712525: certs: publish all in-use public cert data
Before this change, only the default generated CA tls.crt data was published to
the `router-ca` configmap in the `openshift-config-managed` namespace. Change
the behavior to also concatenate any tls.crt data referenced by an
ingresscontroller's default certificate configuration.

This change is necessary because the ingresscontroller is the only way users can
propagate their CA data to kube-controller-manager, which is required for TLS
Route connectivity from pods.