NE-150: Use observedGeneration in admission control
Update status.observedGeneration when processing an ingresscontroller for
admission, and reprocess an ingresscontroller if its generation differs
from its status.observedGeneration in order to ensure that it is
revalidated if it is mutated.

https://jira.coreos.com/browse/NE-150

* pkg/operator/controller/ingress/controller.go (Reconcile): Call
needsReadmission, and reprocess the ingresscontroller if needed.
(admit): Update the ingresscontroller's status.observedGeneration field
when updating its "Admitted" status condition.
(needsReadmission): New function.  Return a Boolean value indicating
whether the given ingresscontroller needs to be reprocessed for admission,
based on its current generation and last observed generation.
* pkg/operator/controller/ingress/status.go (ingressStatusesEqual): Compare
observed generation.