status: Set Degraded using other status conditions
Add a DeploymentDegraded status condition for ingresscontrollers.  Use the
Admitted, DeploymentDegraded, LoadBalancerReady, and DNSReady status
conditions to compute the status of the Degraded status condition.

* pkg/api/v1/types.go (IngressControllerDeploymentDegradedConditionType):
New constant.
* pkg/operator/controller/ingress/controller.go (Reconcile): Check whether
ensureIngressController returns a retryable error, and if so, requeue after
the duration that it specifies.
(ensureIngressController): Use NewMaybeRetryableAggregate to return a
retryable error if the accumulated error values are retryable.
* pkg/operator/controller/ingress/status.go (syncIngressControllerStatus):
Call computeDeploymentDegradedCondition.  Adjust call to
computeIngressDegradedCondition.  Collect error values, and use
NewMaybeRetryableAggregate to return a retryable error if
computeIngressDegradedCondition returns one.
(computeDeploymentDegradedCondition): New function, based on the old
definition of computeIngressDegradedCondition.  Compute and return the
DeploymentDegraded status condition.
(computeIngressDegradedCondition): Delete the deployment parameter, and add
a conditions parameter.  Compute the Degraded status condition from the
ingresscontroller's other status conditions.  Return the status condition
and an error value, which may be a retryable error to indicate that the
status condition should be recomputed after the period specified by the
retryable error value.
* pkg/operator/controller/ingress/status_test.go
(TestComputeIngressDegradedCondition): New test.
* pkg/util/retryableerror/retryableerror.go: New file.
(Error): New type that represents an error for an operation that should be
retried.
(New): New function that returns a retryable error.
(retryableError): New type that wraps an error value and has a time
duration value.
(After): New method that returns the time period after which the operation
that caused a retryable error should be retried.
(NewMaybeRetryableAggregate): New method that given a slice of errors
returns nil if the slice was empty or only contained nil values, a
retryable error if all of the errors were retryable, or else an aggregate
error value.
* pkg/util/retryableerror/retryableerror_test.go: New file.
(TestRetryableError): New test for NewMaybeRetryableAggregate.