Add migration for affinity and deployment strategy
When updating a router deployment, check whether the deployment strategy is
changing; if it is, update the affinity policy first, and only update the
deployment strategy once all pods have the new affinity policy.

This two-step migration is necessary because an old affinity policy may
have an anti-affinity rule that prohibits colocation of pods belonging to
the same ingresscontroller, the new deployment strategy may require surge,
and the presence of pods with the old affinity policy would block a rolling
update using the new deployment strategy.

Follow-up to commit 0f6fd1c67dd719508c33e475ebda3f6e47052501.

This commit fixes bug 1766851.

https://bugzilla.redhat.com/show_bug.cgi?id=1766851

* pkg/operator/controller/ingress/deployment.go (updateRouterDeployment):
If the deployment strategy is changing, make sure to update the affinity
policy first.