Do not wrap errors from syncIngressControllerStatus
An error value from syncIngressControllerStatus may represent a failure to
update status, in which case syncIngressControllerStatus itself adds a
clear error message and wrapping it is therefore unnecessary; or the error
value may represent a retryable error, in which case wrapping the error
value loses the retryable error type and adds an inaccurate error message.

This commit fixes bug 1781948.

https://bugzilla.redhat.com/show_bug.cgi?id=1781948

* pkg/operator/controller/ingress/controller.go (ensureIngressController):
Do not wrap the error value from syncIngressControllerStatus when adding it
to the list of returned error values.
* pkg/util/retryableerror/retryableerror_test.go (TestRetryableError): Add
a test case with a nil error value, and add a nil error value to the "only
retryable errors" test case to verify that nil error values are ignored.