Ignore spec.replicas in deployment hash
When computing the hash for a deployment for use in the pod template spec's
labels and affinity policy, ignore spec.replicas and other fields to which
updates should not trigger a rolling update of the deployment.

This commit fixes bug 1783007.

https://bugzilla.redhat.com/show_bug.cgi?id=1783007

* pkg/operator/controller/ingress/deployment.go (desiredRouterDeployment):
Use the new deploymentTemplateHash function instead of deploymentHash.
(hashableDeployment): Add a parameter to indicate whether to hash fields in
the deployment spec besides the pod template spec.
(deploymentHash): Update call to hashableDeployment.
(deploymentTemplateHash): New function.  Use hashableDeployment.  Return a
hash based on the deployment's metadata and pod template spec.
* pkg/operator/controller/ingress/deployment_test.go
(TestDesiredRouterDeployment): Use deploymentTemplateHash.
(otherToleration): Factor out of TestDeploymentConfigChanged.
(TestDeploymentHash): New test.  Verify that the hashes that deploymentHash and
deploymentTemplateHash return change exactly when expected with respect to
mutations to a deployment.
* test/e2e/operator_test.go (TestIngressControllerScale): Verify that
scaling the deployment does not create a new replicaset.