Bug 1793541: Only initialize DNS providers if DNS config is present
If no DNS configuration is provided, don't try to set up provider clients.

This prevents an operator crash if DNS configuration is not provided on
platforms where managed DNS is expected to be supported (e.g. AWS).

In a followup, the provider configuration can be refactored into the provider
implementations themselves, so this part of the code won't need to know anything
about the provider setup beyond picking the right implementation. Then, it would
be safe to always use the appropriate provider for the platform and let the
provider surface configuration errors if DNS records are actually created to
exercise the provider.