Set affinity for "NodePortService"
Configure the same deployment strategy and affinity policy on the
deployment for an IngressController that uses the "NodePortService"
endpoint publishing strategy type as we do for an IngressController that
uses the "LoadBalancerService" or "Private" strategy type.

Follow-up to commit 3d9a5d3a7e15340cdac37eee3651dadf722a15db.

This commit fixes bug 1801410

https://bugzilla.redhat.com/show_bug.cgi?id=1801410

* pkg/operator/controller/ingress/deployment.go
(desiredRouterDeployment): Configure deployment strategy and affinity
policy the same for "NodePortService" as for "LoadBalancerService" and
"Private".