Tune AWS load balancers to be consistent with other platforms
GCP and Azure both hardcode LB timeouts - set AWS to match the
lower bound of Azure (2 failures @ 5s interval, 2 healthy).
By default AWS is 6 failures @ 10s which is a very long timeout
for the router.

In the future, if these attributes become customizable on GCP and
Azure, they should be consistent with our AWS settings.