Router should deploy with a very long grace period
In order to allow routers to be tuned to shutdown when a variable
length of connections have completed, the grace period needs to be
set fairly high. For now, set it to 1 hour until we determine what
tunable limits we should allow. In general few transactions will
take 1 hour to complete.