Implement logging API
This commit resolves NE-285.

https://issues.redhat.com/browse/NE-285

* pkg/operator/controller/ingress/deployment.go (RouterLogLevelEnvName)
(RouterSyslogAddressEnvName, RouterSyslogFormatEnvName)
(RouterSyslogFacilityEnvName): New constants.
(desiredRouterDeployment): Delete logic for the unsupported-logging
annotation.  Using the new accessLoggingForIngressController function,
check whether spec.logging.access is specified on the ingresscontroller,
and configure logging accordingly.
(accessLoggingForIngressController): New function.  Return the access
logging parameters for the given ingresscontroller if it enables access
logging.
* pkg/operator/controller/ingress/deployment_test.go
(TestDesiredRouterDeployment): Verify that the unsupported-logging
annotation overrides spec.logging.access.  Verify that spec.logging.access
has the desired effect.
(TestInferTLSProfileSpecFromDeployment): Use "logs" as the name of the
syslog sidecar container in order to be consistent with the logging API.
* pkg/operator/controller/ingress/rsyslog_configmap.go
(EnableLoggingAnnotation, haproxyLogLevels): Delete definitions.
(ExtraLoggingEnabled): Delete function.
(desiredRsyslogConfigMap): Use accessLoggingForIngressController to
determine whether a configmap is needed.
* test/e2e/operator_test.go (TestSyslogLogging): New test.  Create a pod
that runs an rsyslog instance and an ingresscontroller that logs to that
rsyslog instance and verify that the rsyslog instance logs an HTTP request.
* test/e2e/operator_test.go (TestContainerLogging): New test.  Create an
ingresscontroller that logs to a sidecar container and verify that it
becomes available.