Use max unavailable 25% or 50%, max surge 25%
Change the parameters for an ingress controller deployment's rolling update
strategy when the ingress controller's endpoint publishing strategy type is
Private, LoadBalancerService, or NodePortService: Set max unavailable to
50% if replicas is less than 4 or 25% if it is equal to or greater than 4
to prevent the deployment from immediately reporting negative availability
when it loses a replica, and set max surge to 25% to allow rollouts to
progress more quickly.

* pkg/operator/controller/ingress/deployment.go (desiredRouterDeployment):
Configure the rolling update strategy with max unavailable set to 25% or
50% and max surge set to 25% if the endpoint publishing strategy type is
Private, LoadBalancerService, or NodePortService.
* pkg/operator/controller/ingress/deployment_test.go
(checkRollingUpdateParams): New test helper.  Check the max unavailable and
max surge rolling update parameters of the given deployment against the
provided expected values.
(TestDesiredRouterDeployment): Use checkRollingUpdateParams to verify that
the rolling update parameters are set appropriately with the
LoadBalancerService endpoint publishing strategy with 2 or 8 replicas and
with the HostNetworkStrategyType endpoint publishing strategy.
(TestDeploymentConfigChanged): Add test case for updating the max
unavailable rolling update parameter.