dns: Avoid publishing already published DNSRecords
When publishing a DNSRecord to a zone, check its status to determine
whether the DNSRecord is already published to that zone, and if it is,
skip publishing it to that zone.

This commit fixes bug 1809354.

https://bugzilla.redhat.com/show_bug.cgi?id=1809354

* pkg/operator/controller/dns/controller.go (Reconcile): Update the
DNSRecord's observed generation.
(publishRecordToZones): Use the observed generation and the new
recordIsAlreadyPublishedToZone function to skip publishing the DNSRecord to
a zone if the DNSRecord has previously been observed and its status
conditions indicate that it is already published to the zone.
(recordIsAlreadyPublishedToZone): New function.  Use the provided
DNSRecord's status to determine whether the DNSRecord is already published
to the provided zone.
* pkg/operator/controller/dns/controller_test.go
(TestRecordIsAlreadyPublishedToZone): New test.