Add configmap for service CA bundle
Create a configmap that is annotated to request injection of the service CA
bundle, and configure ingress controller deployments to mount this
configmap and use it instead of using the service account token to get the
service CA.

See bug 1813894.

https://bugzilla.redhat.com/show_bug.cgi?id=1813894

* assets/router/deployment.yaml: Add a volume and volume mount for the new
configmap, and set the DEFAULT_DESTINATION_CA_PATH environment variable to
configure openshift-router to use the bundle from the configmap.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/ingress/controller.go (ensureIngressController):
Call ensureServiceCAConfigMap once the "openshift-ingress" namespace has
been created and before creating the deployment.
* pkg/operator/controller/ingress/serviceca_configmap.go: New file.
(ensureServiceCAConfigMap): New method.  Ensure the configmap for the
service CA bundle exists.
(desiredServiceCAConfigMap): New function.  Return the desired configmap
for the service CA bundle.
(currentServiceCAConfigMap): New method.  Return the current configmap for
the service CA bundle.
(updateServiceCAConfigMap): New method.  Update the service CA configmap
bundle if the expected annotation is missing.
* pkg/operator/controller/names.go (ServiceCAConfigMapName): New function.
Return the name of the configmap for the service CA bundle.