dns: Improve logging of reconciliation and updates
Improve the logging in the DNS controller and DNS provider implementations
to improve visibility into the controller's operation and to avoid printing
old and possibly misleading status during reconciliation.

* pkg/dns/aws/dns.go (change):
* pkg/dns/azure/dns.go (Ensure, Delete): Log the dnsrecord's spec but not
its status because the controller only updates the status after ensuring or
deleting the dnsrecord.  Log the zone along with the dnsrecord's spec.
* pkg/operator/controller/dns/controller.go (Reconcile): Log when beginning
reconciliation.  When updating the status, log the updated dnsrecord rather
than the old dnsrecord (with the old status), and always log when updating
(not only when the update fails).
(publishRecordToZones): Log when skipping a zone.  Log the dnsrecord's
spec but not its status because the controller only updates the status
after publishRecordToZones returns.