Bug 1862892: Reconcile openshift-ingress-router clusterrole on upgrades
On upgrades we have never needed to handle changes to the
openshift-ingress-router cluster role. This broke 4.5->4.6 upgrades
because the router switched to watching endpointslices in 4.6 and, as
this permission was not granted to the role, it failed to start.

This change now ensures that the cluster role gets updated/reconciled
with the (bindata) values from assets/router/cluster-role.yaml. In
this particular upgrade case it will now add the endpointslices RBAC
rules allowing the router to start.