Bug 1873728: publishRecordToZones: Fix status merge
In publishRecordToZones, when merging the updated status with the input
status in order to produce the output status, first make a copy of the
input status, and use the copy to avoid mutating the input status.

Before this change, publishRecordToZones mutated the input status, and so
when the caller compared the input status with the output status, it would
always find them to be identical.  As a result, the caller never updated
the DNSRecord's status in the API, which caused the DNS controller to
endlessly retry the update and the ingress controller to report
DNSReady=False.

* pkg/operator/controller/dns/controller.go (publishRecordToZones): Make a
copy of the zone statuses, and pass the copy to mergeStatuses to avoid
mutating the original zone statuses that the caller passed in.
* pkg/operator/controller/dns/controller_test.go
(TestPublishRecordToZonesMergesStatus): Verify that publishRecordToZones
does not mutate the status.