Set publishNotReadyAddresses to true on LB services
Set spec.publishNotReadyAddresses to true on the LoadBalancer-type service
that the operator creates for ingress controllers on cloud platforms.

Setting publishNotReadyAddresses causes the endpointslice controller to
publish endpoints of terminating pods.  Such endpoints are published with
endpoints[].conditions.terminating set to false.  This causes the service
proxy to continue to handle traffic as long as the endpoint exists but
report negative responses to health probes if there is no other endpoint,
which all facilitates graceful termination of router pods.

* assets/router/service-cloud.yaml: Add publishNotReadyAddresses: true.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/ingress/load_balancer_service_test.go
(TestLoadBalancerServiceChanged): Add test case to verify that
loadBalancerServiceChanged checks and updates publishNotReadyAddresses as
needed.