aws: use shared credentials file instead of static only
In order to support various styles of authenitcation, mostly STSAsumeRoleUsingWebhookIdentity, the operator should support using the AWS SDK's default
handling of the shared credential file. So previously the operator only supported static credentials from specific keys in the AWS secret, but now the operator
supports reading `credentials` key to load the shared credential file and pass it on as-is to AWS SDK to authenticate based on the contents of that file.

To make the code easier, the operator now converts the previous access_key, secret_key case into a shared credential file using those values and then using
that like if it was provided.

see https://issues.redhat.com/browse/CO-1256 for more details.