Configure operator metrics handler to use localhost
Configure the operator's metrics handler to listen on 127.0.0.1 only.

Before this commit, metrics were exposed on port 60000, without
authentication.  Although the pod spec does not advertise port 60000, the
port was still accessible outside the pod.  After this commit, port 60000
is inaccessible outside the pod, so the only way to get metrics is by
connecting to port 9393 and authenticating with kube-rbac-proxy.

This commit is related to bug 1896923.

https://bugzilla.redhat.com/show_bug.cgi?id=1896923

* cmd/ingress-operator/start.go (NewStartCommand): Configure metrics to
listen on 127.0.0.1:60000.