Annotate manifests for single-node-developer cluster profile
This partially implements phase 1 of https://github.com/openshift/enhancements#482
and does not change behavior. Initially, all cluster-ingress-operator
manifests are included in the single-node-developer cluster profile.
Follow-on PRs may exclude any of these that are not needed in the
profile.

The changes in manifests/00-custom-resource-definition-internal.yaml
and manifests/00-custom-resource-definition.yaml will be overwritten
at make time until the openshift/api module is updated to a version
which has the corresponding changes.