certificate: Conditionally delete generated default cert
pkg/operator/controller/certificate/default_cert.go:

If an Ingress Controller is modified so that the Spec.DefaultCertificate field
is now set, only delete the prior operator generated default certificate if the
secret referenced by Spec.DefaultCertificate exists. This is to correct the bug identified
in BZ 1887441 in which modifying an Ingress Controller to reference a secret that does not
exist via Spec.DefaultCertificate breaks the authentication operator (which relies on the
router-certs secret).

This commit also cleans up the default_cert.go file to align with recent improvements in
the ingress-operator codebase. Namely, the `currentRouterDefaultCertificate` and
`desiredRouterDefaultCertificate` functions are enhanced so that they return an indicator
boolean (ie haveCert and wantCert) in addition to the values previously returned.