certificate-publisher: Check for ic.Spec.DefaultCertificate
pkg/operator/controller/certificate-publisher/publish_certs.go:
Change desiredRouterCertsGlobalSecret to check if an ingress
controller's Spec.DefaultCertificate secret exists in the
parameter secrets slice. If it does not, fall back to the operator
generated default certificate. This makes sense since if
Spec.DefaultCertificate is updated to point to a non-exists secret,
the certificate controller will not delete the operator generated
default certificate.

This commit essentially inlines a more in-depth version of
controller.RouterEffectiveDefaultCertificateSecretName(...) into
desiredRouterCertsGlobalSecret, and is a part of the fix
for BZ 1887441, in which the router certs global secret is
clobbered by the ingress operator when an ingress controller
specifies a default certificate secret that does not exist
(which could be the result of a simple typo or user error).