Support changing ingresscontroller load balancer scope
Add support for changing ingresscontroller load balancer scope. On AWS and
IBM cloud, this requires deleting the existing load balancer service and
recreating it with the desired scope; if the administrator changes the
scope on the ingresscontroller, the operator will report Progressing=True
until the administrator either reverts the change or deletes the service.
If the administrator deletes the service, then the operator will recreate
the service with the desired scope.  On Azure and GCP, it suffices to
update the existing service's annotations.

This commit is based on 982682f10d57041e81fe0983b00610e036286ee1, which was
reverted by commit 61d57137bb88a62ff8d6ea420fd051d0f480d5a2.  Unlike the
original implementation of this feature, the operator does not delete the
service but instead leaves it up to the administrator to do so as needed.

* pkg/operator/controller/ingress/controller.go
(setDefaultPublishingStrategy): Update scope if needed.
(ensureIngressController): Pass infraConfig to syncIngressControllerStatus.
* pkg/operator/controller/ingress/load_balancer_service.go
(managedLoadBalancerServiceAnnotations): ...to here.  Move definition.  Add
keys for annotations for load-balancer scope for platforms that allow
changing the scope without deleting and recreating the service, using the
InternalLBAnnotations and externalLBAnnotations variables to get the
annotation keys for those platforms.
(loadBalancerServiceChanged): Move comment and reword it.
(IsServiceInternal): New function.  Return a Boolean value indicating
whether the provided service is annotated to request an internal
load-balancer
* pkg/operator/controller/ingress/status.go (syncIngressControllerStatus):
Add infraConfig parameter.  Use infraConfig to get the platform type.
Call the new computeIngressProgressingCondition function with the
ingresscontroller, service, and platform to compute a "Progressing" status
condition for the ingresscontroller.
(computeIngressProgressingCondition): New function.  Compute a
"Progressing" status condition.  In particular, this status condition will
indicate if it is necessary to delete the service so that it can be
recreated with the updated scope.
* pkg/operator/controller/ingress/status_test.go
(TestComputeIngressProgressCondition): New test.  Verify that
computeIngressProgressingCondition returns the expected status condition.
* pkg/operator/controller/status/controller.go (Reconcile): Pass
ingresscontrollers to computeOperatorProgressingCondition.
(computeOperatorProgressingCondition): Report Progressing=True on the
clusteroperator if any ingresscontroller reports Progressing=True.
* pkg/operator/controller/status/controller_test.go
(TestComputeOperatorProgressingCondition): Add test case where an
ingresscontroller is progressing.
* test/e2e/operator_test.go (TestScopeChange): New test.  Verify that the
operator performs the appropriate behavior for the platform on which the
test is running when an ingresscontroller's scope is changed from the
default external scope to internal and then back to external.