Refactor computation of platform status
* pkg/operator/controller/ingress/controller.go (Reconcile): Compute the
platform status from the infrastructure config using GetPlatformStatus.
Pass the platform status to ensureIngressController.
(ensureIngressController): Add a parameter for the platform status.  Pass
the platform status instead of the infrastructure config to
ensureLoadBalancerService.
* pkg/operator/controller/ingress/load_balancer_service.go
(ensureLoadBalancerService): Delete the parameter for the infrastructure
config.  Add a parameter for the platform status.  Delete redundant
computation of platform status (since it is now a parameter).  Delete
computation of the PROXY protocol flag (desiredLoadBalancerService can
compute it).
(desiredLoadBalancerService): Delete PROXY protocol parameter.  Instead,
compute the value using IsProxyProtocolNeeded.
* pkg/operator/controller/ingress/load_balancer_service_test.go
(TestDesiredLoadBalancerService): Refactor to be more table-driven.  Change
test cases to specify which annotations are expected to be present, and
which are expected to be absent, in order to simplify the test logic.
(checkServiceHasAnnotation): Fix logic for when expectValue is false.