canary: Fail on all errors in `Reconcile(...)`
pkg/opeator/controller/canary/controller.go:
This commit forces the canary controller to hard-fail on any errors
encountered during the canary controller's reconcile function since none
of these errors are safe to continue on. This fix is to avoid nil
derefernce panics when the controller fails to ensure a resource.