Add edge termination to canary route
This commit adds edge termination to the canary route. If you use an external loadbalancer that redirects all http traffic to https (us in our case) this route will not serve its purpose and the openshift-ingress-operator will be degraded during an upgrade.
Setting it to 'edge' termination should always work.