canary: Skip TLS verification in canary HTTP client
Commit c98416 switched the canary route from a cleartext route
to an edge terminated route. In order for canary probes to succeed, the
canary client needs to skip TLS verification, since the router's default
certificate may be a self-signed certificate. This change is temporary, since
ideally the router's default certificate (or a new certificate for the canary
route) would be added to the canary client's trust bundle.