nodePortServiceChanged: Fix for clusterIPs
Ignore the clusterIPs field when comparing NodePort services to determine
whether an update is required.

Before this commit, the update logic would keep trying to revert the
default value that the API set for the service's clusterIPs field.

The clusterIPs field is new in Kubernetes 1.20 (OpenShift 4.7).

This commit fixes bug 1936030.

https://bugzilla.redhat.com/show_bug.cgi?id=1936030

* pkg/operator/controller/ingress/nodeport_service.go
(nodePortServiceChanged): Ignore the service's clusterIPs field.
* pkg/operator/controller/ingress/nodeport_service_test.go
(TestNodePortServiceChanged): Verify that nodePortServiceChanged returns
false if the only mutation is that the clusterIPs field's value has
changed.