Ingress rollouts should specify minReadySeconds
Ingress components may run directly behind load balancers and thus
need to delay deployment rollout long enough for the load balancers
to see the new process. Without minReadySeconds, the upgrade process
will immediately delete the old pod as soon as the new pod is ready,
which means that the LB may briefly see no pods, which then triggers
on some LB the behavior of including all hosts (generally if all
hosts are unhealthy the LB will allow requests to go to any host
as opposed to failing closed).

In the future we may wish to allow minReadySeconds to be customized
by customers since not all load balancers will take less than 30s
to converge. However, 30s is a reasonable start.