ingressclass: New controller.
Add a controller that creates and manages an ingressclass for each
ingresscontroller.

* manifests/00-cluster-role.yaml: Give the operator access to
ingressclasses.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/ingressclass/controller.go: New file.  Define a
controller that manages ingressclasses for ingresscontrollers.
(controllerName): New const for the new controller's name.
(New): New function.  Create and return a controller that watches
ingresscontrollers and ingressclasses and reconciles them, using the new
ingressClassHasIngressController function and
ingressClassToIngressController method to map events for ingressclasses to
reconcile requests for ingresscontrollers.
(ingressClassHasIngressController): New function.  Given an ingressclass,
return a Boolean value indicating whether it references an
ingresscontroller.
(ingressClassToIngressController): New method.  Given a client object,
check if the object is an ingressclass with an associated
ingresscontroller, and return a slice of reconciliation requests with a
request for any ingresscontroller that is associated with the ingressclass.
(Config): New type.  Store the configuration needed to create an
ingressclass controller.
(reconciler): New type.  Store the state of an ingressclass controller.
(Reconcile): New method.  Handle a reconciliation request for an
ingresscontroller by ensuring that it has the expected ingressclass.
* pkg/operator/controller/ingressclass/ingressclass.go: New file.
(ensureIngressClass): New method.  Ensure the expected ingressclass exists
for the given ingresscontroller, using the new desiredIngressClass function
and currentIngressClass and updateIngressClass methods.
(desiredIngressClass): New function.  Given an ingresscontroller's name,
return a desired ingressclass.
(currentIngressClass): New method.  Given an ingresscontroller's name,
return its current ingressclass if it exists.
(updateIngressClass): New method.  Given current and desired
ingressclasses, update the current ingress class if needed, using the new
ingressclassChanged function.
(ingressclassChanged): New function.  Compare current and expected
ingressclasses to determine if they match, and return an updated
ingressclass if they do not.
* pkg/operator/controller/names.go (IngressClassName): New function.  Given
an ingresscontroller's name, return the name of the corresponding
ingressclass.
* pkg/operator/operator.go (New): Initialize the new controller.
* test/e2e/operator_test.go (TestDefaultIngressClass): New test.  Verify
that the ingressclass controller has created an ingressclass for the
default ingresscontroller.