Add startup probe to the router deployment
Add a startup probe to allow the router up to 2 minutes start up before
sending liveness and readiness probes in case initial synchronization takes
a substantial amount of time.  Using a startup probe instead of specifying
initialDelaySeconds on the liveness and readiness probes allows the kubelet
to start performing liveness and readiness probes quickly if the initial
synchronization is quick.

This commit fixes bug 1943845.

https://bugzilla.redhat.com/show_bug.cgi?id=1943845

* assets/router/deployment.yaml: Add startup probe.  Delete
initialDelaySeconds for the liveness and readiness probes.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/ingress/deployment.go (desiredRouterDeployment):
Configure the startup probe.
(hashableDeployment): Hash the startup probe.
(deploymentConfigChanged): Update the startup probe.
* pkg/operator/controller/ingress/deployment_test.go
(TestDesiredRouterDeployment): Verify that the startup probe is configured
as expected.  Fix an incorrect error message for a readiness probe check.
(TestDeploymentConfigChanged): Split the liveness and readiness test cases
to verify that an update either one alone is detected as a change.  Add a
test case for the startup probe.